Various changes, primarily testing-related (#492)
* Add discussion on conda-forge installation

* Update license year range

* Add note on miniconda scripts

* Update license year range

* Move --level option config to separate file

* Resolve doctest failures

* Resolve mypy complaint

* Switch to Python 3.10 for CI

* Add CI doctest in docs rst files

* No doctest on macos

* Typo fix

* Change --level option usage

* Add pytest-split and remove pytest-runner

* Update submodule

* Attempt to resolve CI pytest failure

* Update submodule

* Change command line argument option style

* Update submodule

* Update installation docs

* Move tifffile back to main requirements

* Bug fix

* Update imageio usage

* Docstring fix

* Switch to imageio v3

* Update submodule

* Update copyright year

* Switch to imageio v3

* Resolve mypy error
