+++ .github/workflows/pytest_latest.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install Python 3
         uses: actions/setup-python@v4
         with:
-          python-version: "3.9"
+          python-version: "3.10"
       - name: Install lastversion
         run: |
           python -m pip install --upgrade pip
