+++ .github/workflows/pytest_ubuntu.yml
@@ -35,7 +35,7 @@ jobs:
             miniforge-version: latest
             activate-environment: test-env
             use-mamba: true
-            python-version: "3.9"
+            python-version: "3.10"
       # Configure conda environment cache
       - name: Set up conda environment cache
         uses: actions/cache@v3
@@ -78,7 +78,7 @@ jobs:
         run: |
           DURATIONS_FILE=$(mktemp)
           bzcat data/pytest/durations_ubuntu.bz2 > $DURATIONS_FILE
-          pytest -x --cov --durations-path $DURATIONS_FILE --splits 5 --group ${{ matrix.group }} --level 2
+          pytest -x --cov --level=2 --durations-path=$DURATIONS_FILE --splits=5 --group=${{ matrix.group }} --pyargs scico
       # Upload coverage data
       - name: Upload coverage
         uses: actions/upload-artifact@v3
@@ -90,16 +90,17 @@ jobs:
         if: matrix.group == 1
         run: |
           pytest --ignore-glob="*test_*.py" --doctest-modules scico
+          pytest --doctest-glob="*.rst" docs
 
   coverage:
     needs: test
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - name: Set up Python 3.9
+      - name: Set up Python 3.10
         uses: actions/setup-python@v4
         with:
-          python-version: "3.9"
+          python-version: "3.10"
       - name: Install deps
         run: |
           python -m pip install --upgrade pip
