+++ .github/workflows/test_examples.yml
@@ -33,7 +33,7 @@ jobs:
             miniforge-version: latest
             activate-environment: test-env
             use-mamba: true
-            python-version: "3.9"
+            python-version: "3.10"
       # Configure conda environment cache
       - name: Set up conda environment cache
         uses: actions/cache@v3
