+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2021-2023, Los Alamos National Laboratory
+Copyright (c) 2021-2024, Los Alamos National Laboratory
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
