+++ README.rst
@@ -38,6 +38,10 @@
    :target: https://pepy.tech/project/scico
    :alt: PyPI download statistics
 
+.. image:: https://img.shields.io/conda/vn/conda-forge/scico.svg
+    :target: https://anaconda.org/conda-forge/scico
+    :alt: Conda Forge Release
+
 .. image:: https://raw.githubusercontent.com/jupyter/design/master/logos/Badges/nbviewer_badge.svg
    :target: https://nbviewer.jupyter.org/github/lanl/scico-data/tree/main/notebooks/index.ipynb
    :alt: View notebooks at nbviewer
@@ -83,5 +87,5 @@ SCICO is distributed as open-source software under a BSD 3-Clause License (see t
 
 LANL open source approval reference C20091.
 
-(c) 2020-2023. Triad National Security, LLC. All rights reserved.
+(c) 2020-2024. Triad National Security, LLC. All rights reserved.
 This program was produced under U.S. Government contract 89233218CNA000001 for Los Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC for the U.S. Department of Energy/National Nuclear Security Administration. All rights in the program are reserved by Triad National Security, LLC, and the U.S. Department of Energy/National Nuclear Security Administration. The Government has granted for itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide license in this material to reproduce, prepare derivative works, distribute copies to the public, perform publicly and display publicly, and to permit others to do so.
