+++ dev_requirements.txt
@@ -1,7 +1,7 @@
 -r requirements.txt
 pylint
 pytest>=7.3.0
-pytest-runner
+pytest-split
 packaging
 pre-commit
 black>=22.3,<23
