+++ requirements.txt
@@ -2,6 +2,7 @@ typing_extensions
 numpy>=1.20.0
 scipy>=1.6.0
 imageio>=2.17
+tifffile
 matplotlib
 jaxlib>=0.4.3,<=0.4.23
 jax>=0.4.3,<=0.4.23
