+++ scico/flax/examples/data_preprocessing.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2022-2023 by SCICO Developers
+# Copyright (C) 2022-2024 by SCICO Developers
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SCICO package. Details of the copyright and
 # user license can be found in the 'LICENSE' file distributed with the
@@ -18,7 +18,7 @@ import numpy as np
 
 import jax.numpy as jnp
 
-import imageio
+import imageio.v3 as iio
 
 from scico import util
 from scico.examples import rgb2gray
@@ -389,7 +389,7 @@ def images_read(path: str, ext: str = "jpg") -> Array:  # pragma: no cover
     slices = []
     shape = None
     for file in sorted(glob.glob(os.path.join(path, "*." + ext))):
-        image = imageio.imread(file)
+        image = iio.imread(file)
         if shape is None:
             shape = image.shape[:2]
         if shape != image.shape[:2]:
